PY/requirements.txt
# Core dependencies
python-dotenv==1.0.0
requests==2.31.0
websockets==12.0  # For real-time communication with server

# Audio processing (client-side)
numpy==1.24.3
sounddevice==0.4.6
PyAudio==0.2.14

# Wake word detection (client-side)
pvporcupine==2.2.1

# Audio visualization (client-side)
tqdm==4.66.1

# Utility packages
pydantic>=2.0.0  # For data validation
